Updates to info.js and resources-home.html
More items added to info.js, also descriptions added to all items.

On the resources-home.html page serveral paragraphs of info were added under "information" most of it was a copy and paste from shared documents, not sure if this belongs here but figured, why not for now. Of course this will likely be changed.